feat: merge view button into project page + reorganize files
