+++ package.json
@@ -26,6 +26,7 @@
     "firebase": "^10.9.0",
     "lucide-react": "^0.360.0",
     "next": "14.1.4",
+    "randomcolor": "^0.6.2",
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^5.0.1",
